Adjust styles and use Overview and Tabs for inventory and communities
