+++ blocks/action-buttons/action-buttons.css
@@ -1,5 +1,5 @@
-.floorplan-links {
-  padding-top: 20px;
+.action-buttons {
+  padding: 30px 0;
   display: flex;
   justify-content: center;
   gap: 10px;
@@ -7,7 +7,7 @@
   a {
     flex: 0;
     font-size: 14px;
-    padding: 4px 25px;
+    padding: 6px 25px;
     border-radius: 3px;
   }
 }
