+++ blocks/overview/overview.css
@@ -1,0 +1,23 @@
+.overview {
+  dl {
+    display: grid;
+    grid-template-columns: 40% 1fr;
+    gap: 4px;
+    margin: 0;
+  }
+
+  dt {
+    font-weight: 600;
+    text-align: right;
+    padding-right: 5px;
+  }
+
+  dd {
+    margin: 0;
+    color: var(--light-blue);
+  }
+
+  dt, dd {
+    line-height: 1.4em;
+  }
+}
