+++ blocks/tabs/tabs.css
@@ -50,3 +50,63 @@
     }
   }
 }
+
+/** for the minimal variant */
+.tabs.minimal {
+  padding-top: 30px;
+  padding-right: 30px;
+
+  .tabs-list {
+    button {
+      padding: 3px 15px;
+      background: none;
+      border: none;
+      color: var(--link-color);
+      p { font-weight: 600; }
+
+      &[aria-selected="true"] {
+        cursor: initial;
+        border-bottom: 4px solid var(--blue);
+      }
+
+      &[aria-selected="false"]:hover,
+      &[aria-selected="false"]:focus {
+        border: none;
+        background-color: initial;
+      }
+    }
+  }
+
+  .tabs-panel {
+    padding-top: 10px;
+    border: none;
+    border-top: 1px solid var(--dark-color);
+    text-align: left;
+
+    a:any-link {
+      font-weight: normal;
+      color: var(--link-color);
+
+      &:hover {
+        color: #2a6496;
+      }
+    }
+
+    & ul {
+      margin: 0;
+
+      & > li {
+        font-weight: bold;
+
+        & ul {
+          margin-bottom: 10px;
+
+          & li {
+            font-weight: normal;
+            color: var(--blue)
+          }
+        }
+      }
+    }
+  }
+}
