+++ templates/blocks/actionbar/actionbar.css
@@ -15,6 +15,13 @@ body[class*="-template"] {
       padding: 4px 2px 0 0;
       margin: 0 5px 20px;
 
+
+      @media (max-width: 767px) {
+        & {
+          transform: scale(0.8);
+        }
+      }
+
       &.dark-blue {
         background: var(--blue);
         border-color: var(--blue);
