+++ templates/inventory/inventory.css
@@ -7,7 +7,8 @@
   }
 
   .directions {
-    margin-bottom: 30px;
+    padding-right: 15px;
+    flex: 1 1 auto;
 
     a {
       color: var(--blue);
@@ -32,12 +33,7 @@
       background: var(--blue);
       mask: url("/icons/directions.svg") no-repeat center;
       margin-left: 8px;
-      top: 1px;
     }
-  }
-
-  .floorplan-links {
-    padding: 30px 0;
   }
 
   .accordion {
@@ -50,96 +46,26 @@
 }
 
 .inventory .pricing-information {
-  display: flex;
+  flex: 0;
 
-  *> {
-    flex: 0 0 50%;
+  h3 {
+    font-size: var(--heading-font-size-xl);
+  }
+
+  h4 {
+    font-size: var(--body-font-size-s);
+    font-weight: 400;
   }
 }
 
-@media (max-width: 991px) {
-  .inventory .pricing-information {
-    flex-direction: column;
-  }
-}
-
-@media (min-width: 992px) {
-  .inventory .pricing-information .cell.border-right {
-    border-right: 1px solid #ccc;
-  }
-}
-
-.inventory .pricing-information .cell {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  >* {
-    margin: 0 5px 5px;
+.inventory .content-wrapper {
+  .button-container {
     display: flex;
+    gap: 15px;
+    padding: 30px 0;
   }
-}
 
-.inventory .price-row {
-  display: flex;
-  align-items: center;
-}
-
-
-/* Cell heading */
-@media (min-width: 1200px) {
-  .inventory .pricing-information .cell> :first-child {
-    height: 30px;
+  button.fancy {
+    flex: 1 1 auto;
   }
-}
-
-/* Cell middle text */
-.inventory .pricing-information .cell> :nth-child(2) {
-  height: 35px;
-  line-height: 1em;
-  flex-direction: column;
-  justify-content: center;
-}
-
-/* Cell Button */
-.inventory .pricing-information .cell> :nth-child(3) {
-  height: 39px;
-  width: 100%;
-  padding: 0 5px 5px;
-  align-content: center;
-}
-
-/* Styling of cell components */
-@media (min-width: 1200px) {
-  .inventory .pricing-information .cell {
-    padding: 0 10px;
-  }
-}
-
-.inventory .pricing-information .cell h3 {
-  display: inline-flex;
-  align-items: baseline;
-  line-height: 1;
-}
-
-.inventory .pricing-information .cell .per-month {
-  font-size: 14px;
-  font-weight: normal;
-}
-
-.inventory .icon {
-  height: 1em;
-  width: 1em;
-  fill: currentcolor;
-  margin-left: 3px;
-}
-
-.inventory button.fancy {
-  width: calc(100%);
-}
-
-.inventory .button-container {
-  display: flex;
-  gap: 15px;
-  padding: 30px 0;
 }
